archive: learn to include submodules in output archive
This commit is a revival of Lars Hjemli's 2009 patch to provide an
option to include submodules in the output of `git archive`.

The `--recurse-submodules` option (named consistently with fetch, clone,
and ls-files) will recursively traverse submodules in the repository and
consider their contents for inclusion in the output archive, subject to
any pathspec filters. Like other commands that have learned
`--recurse-submodules`, submodules that have not been checked out will
not be traversed.

Signed-off-by: Nikhil Benesch <nikhil.benesch@gmail.com>