index-pack: factor out pack/idx finalization
The procedure for moving the ".pack" and the ".idx" files
into place is the same. Rather than repeat the code, let's
factor it into a helper function.

This has the added benefit of clarifying the lifetime of
"final_pack_name". In the original code it is sometimes
redirected to our reusable "name" buffer; after we reuse
that buffer, the contents are nonsense (but nobody looks at
it, so this is not a bug, just a maintenance issue). In the
new code, the pointer modification is contained in the
helper function.

Signed-off-by: Jeff King <peff@peff.net>