tests: create an interactive gdb session with the 'debug' helper
The 'debug' test helper is supposed to facilitate debugging by running
a command of the test suite under gdb.  Unfortunately, its usefulness
is severely limited, because that gdb session is not interactive,
since the test's, and thus gdb's standard input is attached to
/dev/null (for a good reason, see 781f76b15 (test-lib: redirect stdin
of tests, 2011-12-15)).

Re-attach the original standard input in the debug helper, thus
creating an interactive gdb session, which is much much more useful.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>