Inconsistency between git log and git rev-parse for ^HEAD^@
Hello,

the log and rev-parse commands both support the rev^@ syntax which stands for
all parents of rev.  The log command also supports ^rev^@ to exclude all of the
parents of rev, but rev-parse does not.  Should this be fixed?

If so, the following patch would be a start.

Thanks,
Andreas

--

rev-parse: Add support for ^rev^@

Add support for the ^rev^@ syntax to exclude all of the parents of rev.  This
syntax is already supported by git log.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>