stash: make push -p -q --no-keep-index quiet
Currently when using "git stash push -p -q --no-keep-index", the -q flag
is not passed to the git reset which is executed when --no-keep-index is
also passed in.  This means that git stash is somewhat verbose in this
mode even when the -q flag is passed in.  This was always the case since
"git stash save -p" was introduced in dda1f2a5 ("Implement 'git stash
save --patch'", 2009-08-13).

Properly pass the -q flag on to git reset, to make "git stash push -p
-q" as quiet as it should be.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>