git-news: obtain latest news for your favorite VCS
Today when a user is interested in news regarding the Git, their favorite
version control, it is a challenging task to find out what is actually
happening.  If the user is using a Git in a distribution that packaged
it nicely for them, they may find outdated information in their
distribution using the distributions choice of relaying news.
As Git is a fast paced project, this is not the desired behavior for
most users. They rather prefer the latest news.

One could argue that it is OK to expect the user to read the "What's
cooking" emails or even release notes to get a grasp of the latest
development of Git, but these news do not quite catch the lively
atmosphere of the mailing list with all its polite flame wars and
opinionated arguments on why a patch is written the way it ends up in
the release notes.

It is time to fix the root case of this important user facing problem in
a sensible way that doesn't confuse the user.

Invent a new sub command 'git-news', which presents the latest news
in a way capturing all the interesting tid-bits of the mailing list as
well as the surrounding ecosystem.

As we're rushing the solution of this important problem, we did not
consider alternatives, such that we end up solving it in a backwards
compatible, but forwards incompatible way; clearly nobody in the future
wants to add another command that points to different news source. This is
why we can take the generic name 'git-news', instead of using the rather
specialized 'git-rev-news' command that is harder to remember and type.

Signed-off-by: Stefan Beller <sbeller@google.com>