tests: mark tests that fail when the TEST_DIRECTORY is unusual
Mark tests that fail when the TEST_DIRECTORY has a name that's
unusual, i.e. contains control characters, various punctuation etc.

This change is followed by a change that adds the code that picks up
test_fails_on_unusual_directory_names to test-lib.sh, but comes before
that change to not break most of the test suite during bisecting.

This change is the result of:

    for f in $(grep -B3 last_fail_time .prove |grep sh | perl -pe 's/^  (.*?):/$1/');
        do perl -0666 -pi -e 's/^(test_description)/test_fails_on_unusual_directory_names=1\n$1/m' $f
    done

After running the test suite with the aforementioned change to
test-lib.sh. The only exception is t0000-basic.sh, since its testing
the test suite itself its failing 'test --verbose' test needs a change
to the test suite wrapper code to be excluded.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>