convert: fix return value for apply_multi_file_filter() on SIGPIPE
If a pipe write fails and errno is set to EPIPE then we execute the
"error" path. However, as "err" can still be 0 the return value of
apply_multi_file_filter() would indicate success.

Fix this by setting "err" to the value of "errno" in case of a pipe
write failure.

Noticed-by: Ben Peart <benpeart@microsoft.com>
Signed-off-by: Lars Schneider <larsxschneider@gmail.com>