connect.c: handle errors from split_cmdline
On Tue, Apr 11, 2017 at 01:23:32AM +0200, Ævar Arnfjörð Bjarmason wrote:

> There's one segfault in there:
>
> $ ./t5601-clone.sh --root="xtmp.$(perl -e 'print chr 39')" -v -i -d
> [...]
> Cloning into 'ssh-bracket-clone-plink-4'...
> Segmentation fault
> not ok 45 - single quoted plink.exe in GIT_SSH_COMMAND

Here's a fix for that one. I think there are a few other memory
irregularities in that function, too. I'll send another patch in a
minute, but I wanted to get this out in case you were working on it,
too.

-- >8 --
Subject: [PATCH] connect.c: handle errors from split_cmdline

Commit e9d9a8a4d (connect: handle putty/plink also in
GIT_SSH_COMMAND, 2017-01-02) added a call to
split_cmdline(), but checks only for a non-zero return to
see if we got any output. Since the function returns
negative values (and a NULL argv) on error, we end up
dereferencing NULL and segfaulting.

Arguably we could report on the parsing error here, but it's
probably not worth it. This is a best-effort attempt to see
if we are using plink. So we can simply return here with
"no, it wasn't plink" and let the shell actually complain
about the bogus quoting.

Reported-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>