What's cooking in git.git (Apr 2017, #01; Tue, 11)
Junio C Hamano wrote:

> * jk/no-looking-at-dotgit-outside-repo-final (2016-10-26) 1 commit
>   (merged to 'next' on 2017-02-27 at 7373a1b73d)
>  + setup_git_env: avoid blind fall-back to ".git"
>
>  This is the endgame of the topic to avoid blindly falling back to
>  ".git" when the setup sequence said we are _not_ in Git repository.
>  A corner case that happens to work right now may be broken by a
>  call to die("BUG").

There's one caller we missed, in "git apply" when you apply a binary
patch outside any repository.

The patch below fixes it but doesn't have tests.  I'll try to send
out a patch with tests later today.

It's probably also worth making the die("BUG") compile-time
configurable so distros have an easy way out if any similar bugs are
lurking undiscovered.

Thanks,
Jonathan

-- >8 --
Subject: has_sha1_file: don't bother if we are not in a repository

Most callers to this function already require that they are in a
git repository, but there is an exception: "git apply" uses
has_sha1_file to avoid work if the result of applying a binary
patch is already present in the repository. When run outside any
repository, this produces an error:

 fatal: BUG: setup_git_env called without repository

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>